Cleanup bundled runtime entry point
