--- v3/internal/runtime/desktop/@wailsio/runtime/src/log.js
@@ -1,13 +1,0 @@
-/**
- * Logs a message to the console with custom formatting.
- * @param {string} message - The message to be logged.
- * @return {void}
- */
-export function debugLog(message) {
-    // eslint-disable-next-line
-    console.log(
-        '%c wails3 %c ' + message + ' ',
-        'background: #aa0000; color: #fff; border-radius: 3px 0px 0px 3px; padding: 1px; font-size: 0.7rem',
-        'background: #009900; color: #fff; border-radius: 0px 3px 3px 0px; padding: 1px; font-size: 0.7rem'
-    );
-}
